feat: note and octave update based on selected note
