+++ script/data/note.js
@@ -1,50 +1,50 @@
-module.exports=[
+module.exports = [
   {
-    noteName: "c",
-    octave: "4",
+    noteName: "cb",
+    octave: "44",
     duration: "q",
-    domId: "1"
+    domId: "1",
   },
   {
     noteName: "b",
     octave: "4",
     duration: "qr",
-    domId: "2"
+    domId: "2",
   },
   {
     noteName: "b",
     octave: "4",
     duration: "q",
-    domId: "3"
+    domId: "3",
   },
   {
     noteName: "e",
     octave: "5",
     duration: "q",
-    domId: "4"
+    domId: "4",
   },
   {
     noteName: "b",
     octave: "4",
     duration: "q",
-    domId: "1"
+    domId: "1",
   },
   {
     noteName: "b",
     octave: "4",
     duration: "q",
-    domId: "2"
+    domId: "2",
   },
   {
     noteName: "b",
     octave: "4",
     duration: "q",
-    domId: "3"
+    domId: "3",
   },
   {
     noteName: "b",
     octave: "4",
     duration: "q",
-    domId: "4"
+    domId: "4",
   },
-]
+];
